backend/requirements.txt
Flask
Flask-CORS
pymongo
Werkzeug
PyJWT